requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
-e .
dawdreamer==0.8.0
hydra-core==1.3.2
hydra-colorlog==1.2.0
lightning==2.2.3 
matplotlib==3.8.4
optuna==3.5.0
omegaconf==2.3.0
pandas==2.2.2
python-dotenv==0.21.0
pytest==7.4.0
rich==13.3.5
timm==0.9.10 # 0.4.12 might be required for passt 
tqdm==4.65.0
torch==2.2.1
torchaudio==2.2.1
torchopenl3==1.0.1
umap-learn==0.5.6
wandb==0.16.5